Extract task attachment controller slice
