+++ apps/core-api/src/tasks/tasks.controller.ts
@@ -3,16 +3,13 @@ import {
   BadRequestException,
   Controller,
   Delete,
-  ForbiddenException,
   Get,
   Param,
   Patch,
   Post,
   Put,
   Query,
-  UploadedFile,
   UseGuards,
-  UseInterceptors,
   ConflictException,
   NotFoundException,
   Inject,
@@ -29,7 +26,6 @@ import {
   IsOptional,
   IsString,
   IsUUID,
-  MaxLength,
   Max,
   Min,
   ValidateNested,
@@ -51,11 +47,6 @@ import {
 import { SubtaskService } from './subtask.service';
 import { SearchService } from '../search/search.service';
 import { createTaskLifecycleUnitOfWorkFromTx } from './task-lifecycle-prisma.adapter';
-import { FileInterceptor } from '@nestjs/platform-express';
-import { promises as fs } from 'node:fs';
-import { randomBytes, randomUUID } from 'node:crypto';
-import { dirname } from 'node:path';
-import { resolveAttachmentPath } from './attachment-storage';
 import {
   parseRuleDefinition,
   templateDefinition,
@@ -70,7 +61,6 @@ import {
   type ParsedCustomFieldValue,
   type TaskCustomFieldFilter,
 } from '../custom-fields/custom-field.validation';
-import { AttachmentDownloadUrlService } from './attachment-download-url.service';
 import { TaskMentionsService } from './task-mentions.service';
 
 class TaskQuery {
@@ -426,26 +416,6 @@ class PatchDescriptionDto {
   expectedVersion!: number;
 }
 
-class InitiateAttachmentDto {
-  @IsString()
-  @MaxLength(255)
-  fileName!: string;
-
-  @IsString()
-  @MaxLength(120)
-  mimeType!: string;
-
-  @IsInt()
-  @Min(1)
-  @Max(10_000_000)
-  sizeBytes!: number;
-}
-
-class CompleteAttachmentDto {
-  @IsUUID()
-  attachmentId!: string;
-}
-
 class UpsertTaskReminderDto {
   @IsISO8601()
   remindAt!: string;
@@ -495,8 +465,6 @@ class AddDependencyDto {
 
 const MAX_DESCRIPTION_DOC_BYTES = 200_000;
 const MAX_DESCRIPTION_TEXT_LENGTH = 20_000;
-const MAX_IMAGE_UPLOAD_BYTES = 5_000_000;
-const IMAGE_MIME_ALLOWLIST = new Set(['image/png', 'image/jpeg', 'image/webp', 'image/gif']);
 const TIMELINE_LANE_ORDER_BASE_LIMIT = 500;
 const TIMELINE_GROUP_BY_VALUES = ['section', 'assignee'] as const;
 const TIMELINE_VIEW_MODE_VALUES = ['timeline', 'gantt'] as const;
@@ -577,7 +545,6 @@ export class TasksController {
     @Inject(SubtaskService) private readonly subtaskService: SubtaskService,
     @Inject(SearchService) private readonly searchService: SearchService,
     @Inject(NotificationsService) private readonly notifications: NotificationsService,
-    @Inject(AttachmentDownloadUrlService) private readonly attachmentDownloadUrls: AttachmentDownloadUrlService,
     @Inject(TaskMentionsService) private readonly mentions: TaskMentionsService,
   ) {}
 
@@ -1950,239 +1917,6 @@ export class TasksController {
     });
   }
 
-  @Get('tasks/:id/attachments')
-  async listAttachments(
-    @Param('id') taskId: string,
-    @Query('includeDeleted') includeDeletedRaw: string | undefined,
-    @CurrentRequest() req: AppRequest,
-  ) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.VIEWER);
-    const includeDeleted = String(includeDeletedRaw ?? '').toLowerCase() === 'true';
-    const where: Prisma.TaskAttachmentWhereInput = {
-      taskId,
-      completedAt: { not: null },
-      ...(includeDeleted ? {} : { deletedAt: null }),
-    };
-    const attachments = await this.prisma.taskAttachment.findMany({
-      where,
-      orderBy: { createdAt: 'desc' },
-    });
-    return attachments.map((item) => this.serializeAttachment(item));
-  }
-
-  @Post('tasks/:id/attachments/initiate')
-  async initiateAttachment(
-    @Param('id') taskId: string,
-    @Body() body: InitiateAttachmentDto,
-    @CurrentRequest() req: AppRequest,
-  ) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.MEMBER);
-    if (!IMAGE_MIME_ALLOWLIST.has(body.mimeType)) {
-      throw new ConflictException('Unsupported image mime type');
-    }
-    if (body.sizeBytes > MAX_IMAGE_UPLOAD_BYTES) {
-      throw new ConflictException('Image too large');
-    }
-
-    const uploadToken = randomBytes(16).toString('hex');
-    const storageKey = `${taskId}/${randomUUID()}-${this.sanitizeFileName(body.fileName)}`;
-    const attachment = await this.prisma.$transaction(async (tx) => {
-      const created = await tx.taskAttachment.create({
-        data: {
-          taskId,
-          uploaderUserId: req.user.sub,
-          fileName: this.sanitizeFileName(body.fileName),
-          mimeType: body.mimeType,
-          sizeBytes: body.sizeBytes,
-          storageKey,
-          uploadToken,
-        },
-      });
-      await this.domain.appendAuditOutbox({
-        tx,
-        actor: req.user.sub,
-        entityType: 'Task',
-        entityId: taskId,
-        action: 'task.attachment.initiated',
-        afterJson: created,
-        correlationId: req.correlationId,
-        outboxType: 'task.attachment.initiated',
-        payload: { taskId, attachmentId: created.id },
-      });
-      return created;
-    });
-
-    return {
-      attachmentId: attachment.id,
-      uploadUrl: `/attachments/${attachment.id}/upload?token=${uploadToken}`,
-      maxBytes: MAX_IMAGE_UPLOAD_BYTES,
-    };
-  }
-
-  @Post('attachments/:id/upload')
-  @UseInterceptors(
-    FileInterceptor('file', {
-      limits: { fileSize: MAX_IMAGE_UPLOAD_BYTES },
-    }),
-  )
-  async uploadAttachment(
-    @Param('id') id: string,
-    @Query('token') token: string,
-    @UploadedFile() file: { mimetype: string; size: number; buffer: Buffer },
-    @CurrentRequest() req: AppRequest,
-  ) {
-    if (!token) throw new ConflictException('Missing upload token');
-    const attachment = await this.prisma.taskAttachment.findUniqueOrThrow({
-      where: { id },
-      include: { task: true },
-    });
-    await this.domain.requireProjectRole(attachment.task.projectId, req.user.sub, ProjectRole.MEMBER);
-    if (!attachment.uploadToken || attachment.uploadToken !== token) {
-      throw new ForbiddenException('Invalid upload token');
-    }
-    if (!file) throw new ConflictException('Missing file');
-    if (!IMAGE_MIME_ALLOWLIST.has(file.mimetype)) throw new ConflictException('Unsupported image mime type');
-    if (file.size <= 0 || file.size > MAX_IMAGE_UPLOAD_BYTES) throw new ConflictException('Image too large');
-
-    const diskPath = resolveAttachmentPath(attachment.storageKey);
-    await fs.mkdir(dirname(diskPath), { recursive: true });
-    await fs.writeFile(diskPath, file.buffer);
-
-    await this.prisma.taskAttachment.update({
-      where: { id: attachment.id },
-      data: { sizeBytes: file.size, mimeType: file.mimetype },
-    });
-    return { ok: true };
-  }
-
-  @Post('tasks/:id/attachments/complete')
-  async completeAttachment(
-    @Param('id') taskId: string,
-    @Body() body: CompleteAttachmentDto,
-    @CurrentRequest() req: AppRequest,
-  ) {
-    const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
-    await this.domain.requireProjectRole(task.projectId, req.user.sub, ProjectRole.MEMBER);
-    const attachment = await this.prisma.taskAttachment.findUniqueOrThrow({
-      where: { id: body.attachmentId },
-    });
-    if (attachment.taskId !== taskId) throw new ConflictException('Attachment does not belong to task');
-    if (attachment.deletedAt) throw new NotFoundException('Attachment not found');
-
-    const diskPath = resolveAttachmentPath(attachment.storageKey);
-    const stat = await fs.stat(diskPath).catch(() => null);
-    if (!stat) throw new ConflictException('Attachment upload not found');
-
-    const accessToken = randomBytes(16).toString('hex');
-    return this.prisma.$transaction(async (tx) => {
-      const completed = await tx.taskAttachment.update({
-        where: { id: attachment.id },
-        data: { completedAt: new Date(), uploadToken: accessToken, sizeBytes: Number(stat.size) },
-      });
-      await this.domain.appendAuditOutbox({
-        tx,
-        actor: req.user.sub,
-        entityType: 'Task',
-        entityId: taskId,
-        action: 'task.attachment.created',
-        afterJson: completed,
-        correlationId: req.correlationId,
-        outboxType: 'task.attachment.created',
-        payload: { taskId, attachmentId: completed.id },
-      });
-      return this.serializeAttachment(completed);
-    });
-  }
-
-  @Delete('attachments/:id')
-  async deleteAttachment(@Param('id') id: string, @CurrentRequest() req: AppRequest) {
-    const attachment = await this.prisma.taskAttachment.findUniqueOrThrow({
-      where: { id },
-      include: { task: true },
-    });
-    await this.domain.requireProjectRole(attachment.task.projectId, req.user.sub, ProjectRole.MEMBER);
-    if (attachment.deletedAt) return this.serializeAttachment(attachment);
-    return this.prisma.$transaction(async (tx) => {
-      const deleted = await tx.taskAttachment.update({
-        where: { id },
-        data: { deletedAt: new Date() },
-      });
-      await this.domain.appendAuditOutbox({
-        tx,
-        actor: req.user.sub,
-        entityType: 'Task',
-        entityId: attachment.taskId,
-        action: 'task.attachment.deleted',
-        beforeJson: attachment,
-        afterJson: deleted,
-        correlationId: req.correlationId,
-        outboxType: 'task.attachment.deleted',
-        payload: { taskId: attachment.taskId, attachmentId: id },
-      });
-      return this.serializeAttachment(deleted);
-    });
-  }
-
-  @Post('attachments/:id/restore')
-  async restoreAttachment(@Param('id') id: string, @CurrentRequest() req: AppRequest) {
-    const attachment = await this.prisma.taskAttachment.findUniqueOrThrow({
-      where: { id },
-      include: { task: true },
-    });
-    await this.domain.requireProjectRole(attachment.task.projectId, req.user.sub, ProjectRole.MEMBER);
-    if (!attachment.deletedAt) return this.serializeAttachment(attachment);
-    return this.prisma.$transaction(async (tx) => {
-      const restored = await tx.taskAttachment.update({
-        where: { id },
-        data: { deletedAt: null, uploadToken: randomBytes(16).toString('hex') },
-      });
-      await this.domain.appendAuditOutbox({
-        tx,
-        actor: req.user.sub,
-        entityType: 'Task',
-        entityId: attachment.taskId,
-        action: 'task.attachment.restored',
-        beforeJson: attachment,
-        afterJson: restored,
-        correlationId: req.correlationId,
-        outboxType: 'task.attachment.restored',
-        payload: { taskId: attachment.taskId, attachmentId: id },
-      });
-      return this.serializeAttachment(restored);
-    });
-  }
-
-  private serializeAttachment(item: {
-    id: string;
-    taskId: string;
-    uploaderUserId: string;
-    fileName: string;
-    mimeType: string;
-    sizeBytes: number;
-    completedAt: Date | null;
-    createdAt: Date;
-    deletedAt: Date | null;
-    uploadToken: string | null;
-  }) {
-    return {
-      id: item.id,
-      taskId: item.taskId,
-      uploaderUserId: item.uploaderUserId,
-      fileName: item.fileName,
-      mimeType: item.mimeType,
-      sizeBytes: item.sizeBytes,
-      completedAt: item.completedAt,
-      createdAt: item.createdAt,
-      deletedAt: item.deletedAt,
-      url:
-        item.deletedAt || !item.completedAt
-          ? null
-          : this.attachmentDownloadUrls.buildUrl(item),
-    };
-  }
-
   @Get('tasks/:id/reminder')
   async getMyReminder(@Param('id') taskId: string, @CurrentRequest() req: AppRequest) {
     const task = await this.prisma.task.findFirstOrThrow({ where: { id: taskId, deletedAt: null } });
@@ -2766,10 +2500,6 @@ export class TasksController {
     if (descriptionDoc.type !== 'doc' || !Array.isArray(descriptionDoc.content)) {
       throw new ConflictException('descriptionDoc must be a valid ProseMirror doc');
     }
-  }
-
-  private sanitizeFileName(value: string) {
-    return value.replace(/[^a-zA-Z0-9._-]/g, '_').slice(0, 120) || 'upload.bin';
   }
 
   private extractPlainTextFromDoc(node: unknown): string {
